Showing user name at top right & re-direct to base page
When you login with twitter account,
it will re-direct to our happy-hour-restaurants page.

When you are logged in as user,
 it will display Welcome username else it will display Welcome Guest

No logout and Login as Admin link available if not a user.
Also, very small thing but you can go back to the page
by clicking on our website name on top left now.