Showing message on successful login with twitter
When the user successfully logs in with their twitter account,
the pop up appears(which can be removed by closing it with x button)

The empty.txt file is for overwriting existing django message.
Maybe there is a better place to put receiver or ways to making it in to class?